feat(metrics): introduce InsertionDeletionAggregator
In standard attribution methods, the scores are of shape (p, t)
because all targets have the same "p" perturbations.
For Insertion and Deletion metrics, the scores are of shape (p*t, t) because
each target has "p" perturbations that are different for each target.

The idea of this aggregator is to return a tensor of shape (p, t) where for
each target, we only return the corresponding scores.