feat(metrics): support NaN in attributions of generation task
In generation tasks, the attributions of "future" tokens are set to NaN, in
order to avoid considering them in metrics or visualizations. For example,
for a sequence of 6 tokens where 3 were generated (i.e. 3 targets):

attributions = [
* * * NaN NaN NaN
* * *  *  NaN NaN
* * *  *   *  NaN
]

The mask for insertion/deletion metrics consider first the existing values
before NaN to sort the most important tokens to insert/delete.

A unit test is added.