fix: handle NaN values in the visualizations
For generation tasks, the attributions of the tokens that are
not yet used in the generation (future tokens) are now set to nan.

This patch updates the way the min & max values are managed when
computing a normalization, so the nan values are not used.
It also updates the way these attributions are serialized into
JSON to pass them to the js layer (nan are not recognized by the
parser to we need to replace them by None at this step).

Signed-off-by: Frederic Boisnard <frederic.boisnard@irt-saintexupery.com>